backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
psycopg2-binary==2.9.10
sqlalchemy==2.0.39
alembic==1.15.1
python-multipart==0.0.9
pgvector==0.3.6
google-genai
pydantic-settings